v2.2.5 - Send to subfolder - Fix #9
This version allows sending files to sub-folders inside the main shared folder.   
  
The folder link must follow Nextcloud's notation, with `?path=%2FFOLDERNAME` appended to the link.  
  
The easiest way is to open the shared folder from the shared link and then copy the URL.  
  
Example of a valid URL:  
```
https://cloud.mydomain.tld/s/TxWdsNX2Ln3X5kxG?path=%2FMYFOLDER

Fixes #9